requirements.txt
asgiref==3.7.2
Django==4.2.8
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
mysqlclient==2.2.0
Pillow==10.1.0
PyJWT==2.8.0
sqlparse==0.4.4
tzdata==2023.3
